Implementei o Cobertura.cpp
